# swagger.yaml
openapi: 3.0.3
info:
  title: Telegram Clone API
  description: |
    A comprehensive API for a Telegram clone with Signal Protocol encryption, real-time messaging, and advanced group features.
    
    ## Features
    - 🔐 **End-to-End Encryption** using Signal Protocol
    - 💬 **Real-time Messaging** with WebSocket support
    - 👥 **Advanced Group Types** (Channels, Supergroups, Forums)
    - 📁 **File Upload** and media sharing
    - 🔒 **JWT Authentication** with refresh tokens
    - 📱 **Cross-platform** support
    
    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-access-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15 minutes per IP address.
    
    ## WebSocket
    Real-time messaging is available via WebSocket connection at `ws://localhost:3000`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@telegramclone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.telegramclone.com/api/v1
    description: Production server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with Signal Protocol encryption keys
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                summary: Basic registration
                value:
                  username: "john_doe"
                  email: "john@example.com"
                  password: "securepassword123"
                  phoneNumber: "+1234567890"
                  firstName: "John"
                  lastName: "Doe"
              minimal:
                summary: Minimal registration
                value:
                  username: "jane_smith"
                  email: "jane@example.com"
                  password: "securepassword123"
                  phoneNumber: "+1234567891"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user and return access tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              username:
                summary: Login with username
                value:
                  identifier: "john_doe"
                  password: "securepassword123"
              email:
                summary: Login with email
                value:
                  identifier: "john@example.com"
                  password: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Logout user and revoke refresh token
      operationId: logoutUser
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - phoneNumber
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
          description: Username (3-30 characters, alphanumeric and underscore only)
        email:
          type: string
          format: email
          description: Email address
        password:
          type: string
          minLength: 8
          description: Password (minimum 8 characters)
        phoneNumber:
          type: string
          pattern: '^\+[1-9]\d{1,14}$'
          description: Phone number in international format
        firstName:
          type: string
          maxLength: 50
          description: First name
        lastName:
          type: string
          maxLength: 50
          description: Last name
        profilePicture:
          type: string
          format: uri
          description: Profile picture URL
    RegisterResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'
        signalKeys:
          $ref: '#/components/schemas/SignalKeys'
    LoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Username or email
        password:
          type: string
          description: Password
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
        refreshToken:
          type: string
          description: JWT refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'
        signalKeys:
          $ref: '#/components/schemas/SignalKeys'
    RefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: JWT refresh token
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: New JWT access token
        refreshToken:
          type: string
          description: New JWT refresh token
        expiresIn:
          type: integer
          description: Token expiration time in seconds
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Refresh token to revoke
    LogoutResponse:
      type: object
      properties:
        message:
          type: string
          example: "Logged out successfully"

    # User Schemas
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        phoneNumber:
          type: string
          description: Phone number
        firstName:
          type: string
          description: First name
        lastName:
          type: string
          description: Last name
        bio:
          type: string
          description: User bio
        profilePicture:
          type: string
          format: uri
          description: Profile picture URL
        isOnline:
          type: boolean
          description: Online status
        lastSeen:
          type: string
          format: date-time
          description: Last seen timestamp
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
...
    SignalKeys:
      type: object
      properties:
        identityKey:
          type: object
          properties:
            publicKey:
              type: string
              description: Public identity key
            privateKey:
              type: string
              description: Private identity key
        preKeys:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                description: Prekey ID
              publicKey:
                type: string
                description: Public prekey
        signedPreKey:
          type: object
          properties:
            id:
              type: integer
              description: Signed prekey ID
            publicKey:
              type: string
              description: Public signed prekey
            signature:
              type: string
              description: Signature
        registrationId:
          type: integer
          description: Registration ID

    # File Schemas
...
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid input data"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field name
              message:
                type: string
                description: Error message
        code:
          type: string
          example: "VALIDATION_ERROR"
    AuthError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "Invalid or missing authentication token"
        code:
          type: string
          example: "AUTH_ERROR"
    ConflictError:
      type: object
      properties:
        error:
          type: string
          example: "Conflict"
        message:
          type: string
          example: "Resource already exists"
        code:
          type: string
          example: "CONFLICT_ERROR"
...
    ServerError:
      type: object
      properties:
        error:
          type: string
          example: "Internal Server Error"
        message:
          type: string
          example: "An unexpected error occurred"
        code:
          type: string
          example: "SERVER_ERROR"
...